Address remaining PR #756 review comments
- guards.h: clarify SIGNAL_HANDLER_GUARD_OR_DROP() doc to note that guards
  needing cleanup on the drop path (e.g. PerfFdRearmGuard) must be declared
  before the macro too.
- jvmThread_ut.cpp: assert the key-discovery scan finds the right key, and
  fix the resulting flaky test by using a monotonically unique marker
  instead of `this` (whose address gets reused across test instances,
  colliding with never-deleted stale keys from earlier tests).
- jvmThread.h: drop the heavy threadLocalData.h include added only to
  support tickInitWindowIfNeeded(); move that helper to
  threadLocalData.inline.h instead, where ProfiledThread is already
  available, keeping jvmThread.h forward-declaration-only again.
